+++ include/http/RequestParser.hpp
@@ -1,0 +1,19 @@
+#pragma once
+
+namespace http {
+
+	/**
+	 * @class RequestParser
+	 * @brief ...
+	 */
+	class RequestParser {
+		public:
+			RequestParser();
+			~RequestParser();
+			RequestParser(const RequestParser& other);
+			RequestParser& operator=(const RequestParser& rhs);
+
+		private:
+	};
+
+} /* namespace http */
