+++ src/main.cpp
@@ -1,5 +1,6 @@
 #include "config/Parser.hpp"
 #include "core/Reactor.hpp"
+#include "http/Request.hpp"
 
 #include <iostream>
 
@@ -10,6 +11,19 @@ int signal_handler(int signum) {
 int main(int argc, char** const argv) {
 
 	try {
+		std::string dummyReq =
+			"POST /login HTTP/1.1\n"
+			"Host: example.com\n"
+			"Content-Type: application/json\n"
+			"Authorization: Bearer some-jwt-token\n"
+			"User-Agent: MyCustomClient/1.0\n"
+			"Content-Length: 34\n"
+			"\n"
+			"{\"username\": \"user\", \"password\": \"pass123\"}";
+
+
+
+
 
 		std::string configPath = argc > 1 ? argv[1] : "config/configfile_example";
 		config::ConfigFileParser configFileParser;
